% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-cdmDisconnect.R
\name{cdmDisconnect.spark_cdm}
\alias{cdmDisconnect.spark_cdm}
\title{Disconnect the connection of the cdm object}
\usage{
\method{cdmDisconnect}{spark_cdm}(cdm, dropWriteSchema = FALSE, ...)
}
\arguments{
\item{cdm}{cdm reference}

\item{dropWriteSchema}{Whether to drop tables in the writeSchema}

\item{...}{Not used}
}
\description{
Disconnect the connection of the cdm object
}
