% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_omop_cdm_tables.R
\name{create_omop_cdm_tables}
\alias{create_omop_cdm_tables}
\title{Create OMOP CDM tables}
\usage{
create_omop_cdm_tables(con, schema_name, overwrite = FALSE, id_bigint = FALSE)
}
\arguments{
\item{con}{Connection}

\item{schema_name}{Schema in which to create tables}

\item{overwrite}{Whether to overwrite existing tables}

\item{id_bigint}{Whether to use big integers}
}
\value{
OMOP CDM tables created in database
}
\description{
Create OMOP CDM tables
}
