% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOmopTablesOnSpark.R
\name{createOmopTablesOnSpark}
\alias{createOmopTablesOnSpark}
\title{Create OMOP CDM tables}
\usage{
createOmopTablesOnSpark(
  con,
  schemaName,
  cdmVersion = "5.4",
  overwrite = FALSE,
  idBigint = FALSE
)
}
\arguments{
\item{con}{Connection. Must be via odbc (sparklyr connection not supported).}

\item{overwrite}{Whether to overwrite existing tables}

\item{schema_name}{Schema in which to create tables}

\item{cdm_version}{Which version of the OMOP CDM to create. Can be "5.3" or
"5.4"}

\item{id_bigint}{Whether to use big integers}
}
\value{
OMOP CDM tables created in database
}
\description{
Create OMOP CDM tables
}
