% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-dropSourceTable.R
\name{dropSourceTable.spark_cdm}
\alias{dropSourceTable.spark_cdm}
\title{Drop spark tables}
\usage{
\method{dropSourceTable}{spark_cdm}(cdm, name)
}
\arguments{
\item{cdm}{A cdm reference}

\item{name}{The names of the tables to drop. Tidyselect statements can be
used.}
}
\value{
Drops the Spark tables.
}
\description{
Drop Spark tables in the write schema of the connection behind the cdm
reference.
}
