% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropTempTables.R
\name{dropTempTables}
\alias{dropTempTables}
\title{Drop temporary tables.}
\usage{
dropTempTables(src, schema)
}
\arguments{
\item{src}{A \code{cdm_reference} object created with \code{cdmFromSpark()}, a
\code{spark_cdm} source object or a \code{spark_connection} object.}

\item{schema}{A schema, lists are used for schemas, example: schema =
\code{list(catalog = "my_catalog", schema = "my_schema")}}
}
\value{
Invisible TRUE.
}
\description{
Drop temporary tables.
}
