% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-insertTable.R
\name{insertTable.spark_cdm}
\alias{insertTable.spark_cdm}
\title{Insert a table to a cdm object}
\usage{
\method{insertTable}{spark_cdm}(cdm, name, table, overwrite = TRUE, temporary = FALSE)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{name}{The name of the table to insert.}

\item{table}{The table to insert.}

\item{overwrite}{Whether to overwrite an existing table.}

\item{temporary}{If TRUE, a spark dataframe will be written (that will
persist to the end of the current session). If FALSE, a spark table will be
written (which will persist beyond the end of the current session.)}
}
\value{
The cdm reference with the table added.
}
\description{
Insert a local dataframe into the cdm.
}
