% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockSparkCdm.R
\name{mockSparkCdm}
\alias{mockSparkCdm}
\title{creates a cdm reference to local spark OMOP CDM tables}
\usage{
mockSparkCdm(path)
}
\arguments{
\item{path}{A directory for files}
}
\value{
A cdm reference with synthetic data in a local spark connection
}
\description{
creates a cdm reference to local spark OMOP CDM tables
}
\examples{
\dontrun{
mockSparkCdm()
}

}
