% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockSparkCon.R
\name{mockSparkCon}
\alias{mockSparkCon}
\title{creates a local mock database connection}
\usage{
mockSparkCon(
  con = sparklyr::spark_connect(master = "local"),
  databaseName = "GiBleed"
)
}
\arguments{
\item{con}{connection to a database}

\item{databaseName}{The data set name corresponds to the folder with the data set ZIP files}
}
\value{
connection to a db containing Eunomia data
}
\description{
creates a local mock database connection
}
