% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopSparkConnector-package.R
\docType{package}
\name{OmopOnSpark-package}
\alias{OmopOnSpark}
\alias{OmopOnSpark-package}
\title{OmopOnSpark: What the Package Does (One Line, Title Case)}
\description{
What the package does.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopOnSpark/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
