% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compute}
\alias{insertTable}
\alias{cdmTableFromSource}
\alias{listSourceTables}
\alias{dropSourceTable}
\alias{readSourceTable}
\alias{insertCdmTo}
\alias{cdmDisconnect}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{compute}}}

  \item{omopgenerics}{\code{\link[omopgenerics]{cdmDisconnect}}, \code{\link[omopgenerics]{cdmTableFromSource}}, \code{\link[omopgenerics]{dropSourceTable}}, \code{\link[omopgenerics]{insertCdmTo}}, \code{\link[omopgenerics]{insertTable}}, \code{\link[omopgenerics]{listSourceTables}}, \code{\link[omopgenerics]{readSourceTable}}}
}}

