% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkDropTable.R
\name{sparkDropTable}
\alias{sparkDropTable}
\title{Drop a Table from Spark}
\usage{
sparkDropTable(con, schema, tableName)
}
\arguments{
\item{con}{A valid Spark connection object. This is typically created using \code{sparklyr::spark_connect()}.}

\item{schema}{A list specifying the schema details where the table is located. The list may contain:
\itemize{
\item \code{schema}: A string specifying the schema name.
\item \code{catalog} (optional): If included, the function constructs a fully qualified table name with the catalog.
}}

\item{tableName}{A string specifying the name of the table to drop.}
}
\value{
The result of the \verb{DROP TABLE} query execution as returned by \code{DBI::dbGetQuery()}.
}
\description{
This function drops a specified table from a schema in a Spark connection, if it exists.
}
\examples{
\dontrun{
# Establish a Spark connection
con <- sparklyr::spark_connect(master = ...)

# Drop a table from the default schema
sparkDropTable(con, schema = list("schema" = "default"), tableName = "my_table")

# Disconnect from Spark
sparklyr::spark_disconnect(con)
}

}
