% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkListTables.R
\name{sparkListTables}
\alias{sparkListTables}
\title{This function lists the tables available in a specified schema within a Spark connection.}
\usage{
sparkListTables(con, schema)
}
\arguments{
\item{con}{A valid Spark connection object. This is typically created using \code{sparklyr::spark_connect()}.}

\item{schema}{A list specifying the schema and the catalog where the tables are located}
}
\value{
A character vector of table names available in the specified catalog and schema within the Spark connection.
}
\description{
This function lists the tables available in a specified schema within a Spark connection.
}
\examples{
\dontrun{
# Establish a Spark connection
con <- sparklyr::spark_connect(...)

# List tables in the default schema
tables <- sparkListTables(con)

# List tables in a specific schema
tables_in_schema <- sparkListTables(con, schema = list(catalog = "cat", schema = "my_schema"))

# Disconnect from Spark
sparklyr::spark_disconnect(con)
}
}
