% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkReadTables.R
\name{sparkReadTable}
\alias{sparkReadTable}
\title{This function reads a table from a specified schema and table name in a Spark connection.}
\usage{
sparkReadTable(con, schema, tableName)
}
\arguments{
\item{con}{A valid Spark connection object. This is typically created using \code{sparklyr::spark_connect()}.}

\item{schema}{A list specifying the schema and the catalog where the table is located and any prefix}

\item{tableName}{A string specifying the name of the table to read from the Spark connection.}
}
\value{
A \code{tbl} object from the \code{dplyr} package, representing the Spark table specified by the \code{schema} and \code{tableName}.
}
\description{
This function reads a table from a specified schema and table name in a Spark connection.
}
