% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkSource.R
\name{sparkSource}
\alias{sparkSource}
\title{Create a spark source object.}
\usage{
sparkSource(con, cdmSchema, writeSchema, writePrefix = NULL)
}
\arguments{
\item{con}{A connection to a spark database.}

\item{cdmSchema}{A schema for the OMOP CDM core tables.}

\item{writeSchema}{A write schema with writing permissions.}

\item{writePrefix}{A write prefix used whenever writing tables in the write
schema.}
}
\value{
A spark_cdm object.
}
\description{
Create a spark source object.
}
\examples{
\dontrun{
con <- sparklyr::spark_connect(master = "local")
sparkSource(con)
}

}
