% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOmopTablesOnSpark.R
\name{createOmopTablesOnSpark}
\alias{createOmopTablesOnSpark}
\title{Create OMOP CDM tables}
\usage{
createOmopTablesOnSpark(
  con,
  schemaName,
  cdmVersion = "5.4",
  overwrite = FALSE,
  bigInt = FALSE
)
}
\arguments{
\item{con}{Connection to a Spark database.}

\item{schemaName}{Schema in which to create tables.}

\item{cdmVersion}{Which version of the OMOP CDM to create. Can be "5.3" or
"5.4".}

\item{overwrite}{Whether to overwrite existing tables.}

\item{bigInt}{Whether to use big integers for person identifier (person_id
or subject_id)}
}
\value{
OMOP CDM tables created in database
}
\description{
Create OMOP CDM tables
}
