% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockSparkCdm.R
\name{mockSparkCon}
\alias{mockSparkCon}
\title{creates a local mock database connection}
\usage{
mockSparkCon(
  con = sparklyr::spark_connect(master = "local"),
  databaseName = "GiBleed",
  cdmSchema = list(),
  writeSchema = list()
)
}
\arguments{
\item{con}{A spark connection using \code{sparklyr::spark_connect()}.}

\item{databaseName}{Name of the dataset to download.}

\item{cdmSchema}{Schema to include omop standard tables. Schema will be
created if it does not exist.}

\item{writeSchema}{Schema to include cohort tables. Schema will be created if
it does not exist.}
}
\value{
Connection to the spark data set with Eunomia loaded.
}
\description{
creates a local mock database connection
}
\examples{
\dontrun{
con <- mockSparkCon()
}

}
