% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkSource.R
\name{sparkSource}
\alias{sparkSource}
\title{Create a spark source object.}
\usage{
sparkSource(con, writeSchema = NULL)
}
\arguments{
\item{con}{A connection to a spark database.}

\item{writeSchema}{A write schema with writing permissions. Use NULL to use
temp tables.}
}
\value{
A spark_cdm object.
}
\description{
Create a spark source object.
}
\examples{
\dontrun{
con <- sparklyr::spark_connect(master = "local")
sparkSource(con)
}

}
